Level 3 - Ex.5: Modify the previous exercise to use routes and use them in the access button.
