+++ my-app/src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import Welcome from './components/Welcome';
+
 
 ReactDOM.render(
-  <Welcome/>,
+  <App />,
   document.getElementById('root')
 );
